--- a/hail/src/main/scala/is/hail/expr/ir/Exists.scala
+++ b/hail/src/main/scala/is/hail/expr/ir/Exists.scala
@@ -31,16 +31,54 @@ object Exists {
 }
 
 object Forall {
-  def apply(node: BaseIR, visitor: BaseIR => Boolean): Boolean =
-    !Exists(node, n => !visitor(n))
+  def apply(node: IR, visitor: IR => Boolean): Boolean =
+    !Exists.inIR(node, n => !visitor(n))
+}
+
+object IsScanResult {
+  def apply(root: IR): Boolean = root match {
+    case _: ApplyScanOp => true
+    case AggFilter(_, _, isScan) => isScan
+    case AggExplode(_, _, _, isScan) => isScan
+    case AggGroupBy(_, _, isScan) => isScan
+    case AggArrayPerElement(_, _, _, _, _, isScan) => isScan
+    case _ => false
+  }
+}
+
+object IsAggResult {
+  def apply(root: IR): Boolean = root match {
+    case _: ApplyAggOp => true
+    case AggFilter(_, _, isScan) => !isScan
+    case AggExplode(_, _, _, isScan) => !isScan
+    case AggGroupBy(_, _, isScan) => !isScan
+    case AggArrayPerElement(_, _, _, _, _, isScan) => !isScan
+    case _ => false
+  }
 }
 
 object ContainsAgg {
-  def apply(root: BaseIR): Boolean = Exists(root, _.isInstanceOf[ApplyAggOp])
+  def apply(root: IR): Boolean = IsAggResult(root) || (root match {
+    case _: TableAggregate => false
+    case _: MatrixAggregate => false
+    case ArrayAgg(a, _, _) => ContainsAgg(a)
+    case _ => root.children.exists {
+      case child: IR => ContainsAgg(child)
+      case _ => false
+    }
+  })
 }
 
 object ContainsScan {
-  def apply(root: BaseIR): Boolean = Exists(root, _.isInstanceOf[ApplyScanOp])
+  def apply(root: IR): Boolean = IsScanResult(root) || (root match {
+    case _: TableAggregate => false
+    case _: MatrixAggregate => false
+    case ArrayAggScan(a, _, _) => ContainsScan(a)
+    case _ => root.children.exists {
+      case child: IR => ContainsScan(child)
+      case _ => false
+    }
+  })
 }
 
 object Extract {