* redesign LowerMatrixIR to pass through mutable state for relational lets
 * switch several functions on Python frontend to define functions; this
   is to manage the size of generated code