@tpoterba 2019-05-28
This is a big PR, but the bulk of the changed lines come from indentation changes. I'll annotate various pieces to help the review process along. I'll also think about how to break it into smaller pieces.