@tpoterba 2019-05-28
Timings:

Master:
```
running sample_qc...
    run 1 took 27.46s
    run 2 took 24.08s
    run 3 took 23.59s
    Mean, Median: 25.04s, 24.08s

running variant_and_sample_qc...
    run 1 took 54.03s
    run 2 took 45.70s
    run 3 took 45.14s
    Mean, Median: 48.29s, 45.70s
```

PR:
```
running sample_qc...
    run 1 took 18.89s
    run 2 took 17.22s
    run 3 took 16.28s
    Mean, Median: 17.47s, 17.22s

running variant_and_sample_qc...
    run 1 took 38.78s
    run 2 took 39.05s
    run 3 took 35.66s
    Mean, Median: 37.83s, 38.78s
```