@tpoterba 2019-05-28
I removed a couple of unnecessary changes to be added later. This is conceptually small enough for review, even though the diff is still oddly large dude to the LowerMatrixIR change.