@tpoterba 2019-05-29
d'oh -- there's still no lowering rule for `MatrixMultiWrite`. Will PR that separately.