@tpoterba 2019-05-29
MatrixMultiWrite lowering is in.